+++ src/container/MyCouponStack.tsx
@@ -4,7 +4,7 @@ import { Text } from 'react-native-elements'
 
 const MyCoupon = () => {
     return (
-        <View style={{flex: 1, justifyContent: "center", alignItems: "center"}}>
+        <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
             <Text h1>텅</Text>
         </View>
     )
