+++ src/navigation/MyPageStack.tsx
@@ -1,7 +1,6 @@
 import { createStackNavigator } from '@react-navigation/stack';
 import React from 'react'
 import MyPage from '../components/MyPage';
-import MyCoupon from '../components/MyPage/MyCoupon';
 
 const Stack = createStackNavigator();
 const MyPageStack = () => {
@@ -9,10 +8,6 @@ const MyPageStack = () => {
     return (
         <Stack.Navigator >
             <Stack.Screen name="MyPage" component={MyPage} options={{ headerShown: false }}/>
-            <Stack.Screen 
-                name="MyCoupon" 
-                component={MyCoupon} 
-                options={{ headerTitle: "내 쿠폰", headerBackTitleVisible: false}} />
         </Stack.Navigator>
     )
 }
