+++ src/util/index.ts
@@ -1,32 +1,13 @@
-import { Coupon, PinPoint } from "@types";
+import { AuthContext, useAuthContext, AuthContextProvider } from "./Auth";
+export { AuthContext, useAuthContext, AuthContextProvider }
 
-export const isJsonString = (str: string) => {
-  try {
-    JSON.parse(str);
-  } catch (e) {
-    return false;
-  }
-  return true;
-};
+import { LoadingContext, useLoadingContext, LoadingContextProvider } from "./Loading";
+export { LoadingContext, useLoadingContext, LoadingContextProvider }
 
-export const isBlank = (v: string[]) => {
-  return v.some(s => s.replace(" ", "") === "")
-}
+import perventGoBack from "./perventGoBack";
+export { perventGoBack }
 
-const isEqualStringArray = (a: string[], b: string[]) => {
-  if (a.length !== b.length)
-    return false;
-
-  return a.find((item, i) => item !== b[i]) === undefined;
-}
-
-export const isEditPinPoint = (a: PinPoint, b: PinPoint) => {
-  const aq = a.quiz, bq = b.quiz;
-  return a.name !== b.name || a.description !== b.description || a.latitude !== b.latitude
-    || aq.answer !== bq.answer || aq.text !== bq.answer || aq.type !== bq.type
-    || !isEqualStringArray(aq.choices, bq.choices) || !isEqualStringArray(a.imgs, b.imgs);
-}
-export const isEditCoupon = (a: Coupon, b: Coupon) => {
-  return a.name !== b.name || a.description !== b.description
-    || a.endDate !== b.endDate || a.limit !== b.limit || !isEqualStringArray(a.imgs, b.imgs)
+import { isBlank, isEditCoupon, isEditPinPoint, isJsonString } from './util'
+export const $$ = {
+  isBlank, isEditCoupon, isEditPinPoint, isJsonString
 }
